Rust operations under transaction now release the GIL (#50)
Rust operations, like cp, now release the GIL, allowing the code to be
far more asynchronous. Back-of-hand tests indicate a significant
speedup.

The key changes are:
New macro rust_async! in lib.rs. This macro will release the GIL when
called. It also has much improved error handling over the previous
versions. Now the GIL is released on all calls into reading, writing,
etc.

Transaction mechanics changed. The previous block to releasing the GIL
was that all the file write objects had a reference count that they
incremented and decremented in the transaction object. However, this was
not done with any sort of locking. This meant that it was easy for a
transaction to close between when a write file was created and when it
incremented the counter, causing an error.

Now, the transaction operations (e.g. open_for_write) are accessed only
through a unique transaction token, which is essentially a
Arc<Transaction> to the transaction object. When this is dropped, or
explicitly closed (allowing errors to propagate nicely), it gives the
token back to the transaction. When all tokens are returned, the
transaction is closed.

In Rust, there are now two classes exposed to python. PyWriteTransaction
is the transaction handle, and contains a pointer to the internal
transaction class. It can give out new tokens to the current transaction
object. It can also atomically set the current transaction to be
committed or closed once all the current tokens are returned, and
replace the current one with a new instance from which any future tokens
will be issued.

The second class is the PyWriteTransactionAccessToken, which has all the
open_for_write and other modifying methods for interacting with the
transaction.

This mechanic then allows the python end to be vastly simplified, a
layer of locking to be removed, and only one commit transaction class is
needed on that end. Also, all the explicit reference counting semantics
in python are removed.

Also adds an integration test that upload a large batch of files.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>
Co-authored-by: Di <di@xetdata.com>