Lower transaction limits.
Currently, the transaction limits allow 2048 files per commit, and 8
commits to be in flight at once.  However, analysis of the atomic_commit
function, called to complete a transaction, show that this operation
takes roughly 100-150ms per file present in the transaction.  This
results in the client-side piling up transactions on the server, with
the commits later in the queue not completing before the load balancer
times out.  Reducing these limits is a temporary fix to throttle the
number of commits while still allowing CAS and other file uploads to
proceed in parallel.