Build scripts for development and distribution of pyxet / xet cli (#132)
Build scripts for easier development and distribution of pyxet.

This PR adds in several build scripts to aid in the development and
distribution of pyxet. In particular, it enables the distribution of a
xet standalone binary along with the git-xet binary.

Scripts added (run in python/pyxet):

- sourcing `./develop.sh` now sets up the virtual environment if it is
not already set up.
- `./scripts/build_wheel.sh` builds the wheel in a controlled virtual
environment. Works on all platforms.
- `./scripts/build_standalone_cli.sh` builds the standalone xet cli.
Works on all platforms.
- `./scripts/set_libxet.sh` controls which version of xet-core/libxet
the pyxet system depends on. Switches it between specific commit, path,
or the default remote.

All scripts work on all platforms.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>